Makes c3tk a first-class citizen
This addresses #9

In order to make things play a little more nicely on the user
side of the quation, `c3tk install` now installs only the c3tk
script itself as /usr/local/bin/c3tk by default.

This also changes the default behavior of `c3tk add` such that
the symlinks are created within the user's ~/.config/c3tk/bin,
which is hopefully added to their path per the new instructions
that are output by `c3tk install`.

We've also added a `c3th paths` command to make the path update
a bit easier. It inludes the user's c3tk config bindir,
/opt/c3tk/bin (which isn't really a thing yet), and
/usr/local/bin (just for the sake of completeness).

So as to make sure that the user's ~/.config/c3tk doesn't end up
root-owned, we've removed the `install_cmds` function call from
`c3tk install`, and as this is the only place that it was used,
we've also removed the function altogether.

Finally, to make sure things are on the up-and-up, all commands
except for `c3tk install` now create the ~/.config/c3tk tree if
it does not exist.